PMCVIP-831 - Move asset chooser code into it's own class
The main plugin file has become quite large, and serving many purposes. We can split it up.
