+++ includes/class-asset-chooser.php
@@ -1,0 +1,316 @@
+<?php
+
+namespace Webdam;
+
+/**
+ * Class Asset_Chooser
+ * @package Webdam
+ */
+class Asset_Chooser {
+
+	/**
+	 * @var Used to store an internal reference for the class
+	 */
+	private static $_instance;
+
+	/**
+	 * @return Asset_Chooser object instance
+	 */
+	static function I() {
+
+		if ( empty( static::$_instance ) ){
+
+			self::$_instance = new self();
+		}
+
+		// Return the single/cached instance of the class
+		return self::$_instance;
+	}
+
+	/**
+	 * Handles registering hooks that initialize this plugin.
+	 */
+	protected function __construct() {
+
+		add_action( 'wp_enqueue_scripts', array( $this, 'action_wp_enqueue_scripts' ) );
+
+		// Handle sideloading images from WebDAM
+		add_action( 'wp_ajax_pmc-webdam-sideload-image', array( $this, 'handle_ajax_image_sideload' ) );
+
+		//load up plugin functionality only if domain path is saved in options
+		if ( \get_webdam_settings() ) {
+
+			add_filter( 'mce_external_plugins', array( $this, 'mce_external_plugins' ) );
+			add_filter( 'mce_buttons', array( $this, 'mce_add_button' ) );
+
+			// Load admin variable for the domain in the plugin
+			add_action( 'admin_enqueue_scripts', array( $this, 'plugin_load_plugin_vars' ) );
+		}
+	}
+
+	/**
+	 * Enqueue any scripts or styles
+	 */
+	public function action_wp_enqueue_scripts() {
+
+		if ( is_admin() ) {
+
+		} else {
+
+			// Enqueue the webdam imported asset CSS
+			wp_enqueue_style(
+				'webdam-imported-asset',
+				WEBDAM_PLUGIN_URL . 'assets/webdam-imported-asset.css',
+				array(),
+				false,
+				'screen'
+			);
+		}
+	}
+
+	/**
+	 * Enqueues the JS which loads the domain name
+	 *
+	 * @todo localize vars
+	 * @todo Move status markup into _ template
+	 */
+	public function plugin_load_plugin_vars() {
+
+		global $post;
+
+		$screen = get_current_screen();
+
+		// Only output the following <script> on edit/new post screens
+		if ( 'post' !== $screen->base ) {
+			return;
+		}
+
+		$settings = get_webdam_settings();
+
+		$domain_path = $settings['webdam_account_domain'];
+
+		if ( false === strpos( $domain_path, '://' ) ) {
+			$domain_path = 'http://' . $domain_path;
+		}
+
+		wp_enqueue_style(
+			'webdam-chooser-styles',
+			WEBDAM_PLUGIN_URL . 'assets/assetchooser.css',
+			array(),
+			false,
+			'screen'
+		);
+		?>
+
+		<div class="webdam-asset-chooser-status">
+			<div class="working">
+				<?php esc_html_e( 'Importing your WebDAM selection..', 'PMC' ); ?>
+				<img src="/wp-includes/js/thickbox/loadingAnimation.gif" alt="Waiting" />
+			</div>
+			<div class="done"></div>
+		</div>
+		<script type="text/javascript">
+			var webdam_sideload_nonce = <?php echo wp_json_encode( wp_create_nonce( 'webdam_sideload_image' ) ); ?>;
+			var post_id = <?php echo wp_json_encode( $post->ID ); ?>;
+			var asset_chooser_domain = <?php echo wp_json_encode( $domain_path ); ?>;
+		</script>
+		<?php
+	}
+
+	/**
+	 * Initialize TinyMCE table plugin and custom TinyMCE plugin
+	 *
+	 * @param array $plugin_array Array of TinyMCE plugins
+	 * @return array Array of TinyMCE plugins
+	 */
+	public function mce_external_plugins( $plugin_array ) {
+		$plugin_array['webdam_asset_chooser'] = WEBDAM_PLUGIN_URL . 'assets/assetchooser-loader.js';
+		return $plugin_array;
+	}
+
+	/**
+	 * Add TinyMCE table control buttons
+	 *
+	 * @param array $buttons Buttons for the second row
+	 * @return array Buttons for the second row
+	 */
+	public function mce_add_button( $buttons ) {
+		array_push( $buttons, "separator", 'btnWebDAMAssetChooser' );
+        return $buttons;
+	}
+
+	/**
+	 * Sideload the remote WebDAMN image into WP's media library
+	 *
+	 * This is executed over AJAX from client-side when an image is chosen
+	 * in the WebDAM interface.
+	 */
+	public function handle_ajax_image_sideload() {
+
+		// Verify doing ajax
+		if ( ! defined( 'DOING_AJAX' ) && ! DOING_AJAX ) {
+			return;
+		}
+
+		// Verify our nonce to ensure safe origin
+		check_ajax_referer( 'webdam_sideload_image', 'nonce' );
+
+		// Verify we've got the data we need to proceed
+		if ( empty( $_POST['post_id'] ) ) {
+			wp_send_json_error( array( 'No post ID provided.' ) );
+		}
+
+		if ( empty( $_POST['webdam_asset_url'] ) ) {
+			wp_send_json_error( array( 'No image source provided.' ) );
+		}
+
+		// Sanitize our input
+		$post_id          = (int) $_POST['post_id'];
+		$webdam_asset_id  = (int) $_POST['webdam_asset_id'];
+		$webdam_asset_url = esc_url_raw( $_POST['webdam_asset_url'] );
+		$webdam_asset_filename = sanitize_file_name( $_POST['webdam_asset_filename'] );
+
+		// Adjust the remote image url so we receive the largest image possible
+		$webdam_asset_url = str_replace( 'md_', '1280_', $webdam_asset_url );
+
+		// Hook into add_attachment so we can obtain the sideloaded image ID
+		// media_sideload_image does not return the ID, which sucks.
+		add_action( 'add_attachment', array( $this, 'add_attachment' ), 10, 1 );
+
+		// Sideload the image into WP
+		$local_image_url  = media_sideload_image( $webdam_asset_url, $post_id, '', 'src' );
+
+		// Grab the sideloaded image ID we just set via the
+		// add_attachment actionm hook
+		$attachment_id = get_post_meta( $post_id, 'webdam_attachment_id_tmp', true );
+
+		// We don't need this any longer—let's ditch it.
+		delete_post_meta( $post_id, 'webdam_attachment_id_tmp' );
+
+		// Grab the current image metadata
+		$wordpress_image_meta = wp_get_attachment_metadata( $attachment_id );
+
+		// Fetch metadata for the image
+		// Some images contain embeded metadata, but that is unreliable
+		// and often not present. We could create code to check existing data,
+		// and fetch what's needed, but the likelihood of images with data
+		// is slim, and depends on the photographer.
+		$webdam_image_meta = \get_webdam_asset_metadata( $webdam_asset_id );
+
+		// Set the initial alttext
+		$post_alttext = '';
+
+		if ( ! empty( $wordpress_image_meta['image_data']['title'] ) ) {
+			$post_alttext = $wordpress_image_meta['image_data']['title'];
+		}
+
+		if ( false !== $webdam_image_meta ) {
+
+			$post_title = $post_content = $post_excerpt = $post_credit = '';
+
+			if ( ! empty( $webdam_image_meta->headline ) ) {
+				$post_title = $webdam_image_meta->headline;
+				$post_alttext = $webdam_image_meta->headline;
+			}
+
+			if ( ! empty( $webdam_image_meta->caption ) ) {
+				$post_content = $webdam_image_meta->caption;
+				$post_excerpt = $webdam_image_meta->caption;
+			}
+
+			if ( ! empty( $webdam_image_meta->byline ) ) {
+				$post_credit = $webdam_image_meta->byline;
+			}
+
+			// Set the attachment post attributes
+			wp_update_post( array(
+				'ID'           => $attachment_id,
+				'post_title'   => $post_title,
+				'post_content' => $post_content,
+				'post_excerpt' => $post_excerpt,
+			) );
+
+			// Set the attachment post meta values
+			$attachment_post_metas = array(
+				'_wp_attachment_image_alt' => $post_alttext,
+				'_image_credit'            => $post_credit,
+				'_webdam_asset_id'         => $webdam_asset_id,
+				'_webdam_asset_filename'   => $webdam_asset_filename,
+			);
+
+			foreach ( $attachment_post_metas as $meta_key => $meta_value ) {
+				update_post_meta( $attachment_id, $meta_key, $meta_value );
+			}
+
+			// Merge the existing metadata (that WP found embedded within the image)
+			// with the metadata from the WebDAM API
+			if ( ! empty( $wordpress_image_meta['image_data'] ) && is_array( $wordpress_image_meta['image_data'] ) ){
+
+				$wordpress_image_meta['image_data'] = array_merge( $wordpress_image_meta['image_data'], (array) $webdam_image_meta );
+
+			} else {
+
+				$wordpress_image_meta['image_data'] = (array) $webdam_image_meta;
+
+			}
+
+			// Update the metadata stored for the image by WordPress
+			wp_update_attachment_metadata(
+				$attachment_id,
+				$wordpress_image_meta
+			);
+		}
+
+		// Return the local image url on success
+		// ..error message on failure
+		if ( is_wp_error( $local_image_url ) ) {
+
+			wp_send_json_error( array( 'Unable to sideload image.' ) );
+
+		} else {
+
+			wp_send_json_success( array( 'url' => $local_image_url, 'alttext' => $post_alttext ) );
+
+		}
+	}
+
+	/**
+	 * Helper to obtain sideloaded image ID
+	 *
+	 * We add this hook before calling media_sideload_image,
+	 * and remove it immediately afterwards. This allows us to
+	 * capture the newly sideloaded attachment ID. In this context
+	 * we can obtain the post_parent and use that to set post meta
+	 * on the parent post, which contains the attachment ID.
+	 *
+	 * It's a little hacky, but by doing so we can call get_post_meta
+	 * after calling media_sideload_image to obtain the new attachment ID
+	 *
+	 * @internal Called via add_attachment action hook
+	 *
+	 * @param $attachment_id The ID of the newly inserted attachment image
+	 *
+	 * @return null
+	 */
+	public function add_attachment( $attachment_id ) {
+
+		// Remove this hook callback so it doesn't fire again
+		// We only want this to fire once, right when we're sideloading
+		// the image into WP.
+		remove_action( 'add_attachment', array( $this, 'add_attachment' ), 10, 1 );
+
+		// Fetch the attachment's post so we may obtain it's parent ID
+		// When we call media_sideload_image we specify the original post's ID
+		// so that the attachment will be attached to the post.
+		$attachment = get_post( $attachment_id );
+
+		// Set temporary post meta on the parent post so we may obtain the
+		// attachment id via get_post_meta immediately after calling
+		// media_sideload_image()
+		add_post_meta( $attachment->post_parent, 'webdam_attachment_id_tmp', $attachment_id );
+	}
+}
+
+Asset_Chooser::I();
+
+//EOF
